organize queries and actions a little bit
