+++ packages/evmos-wallet/src/server/fetch-token-prices.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from "vitest";
-import { fetchTokenPrices } from "./fetch-token-prices";
+import { fetchTokenPrices } from "./fetch-token-prices.server";
 
 describe("fetch-token-prices", ({}) => {
   test("should fetch token prices", async ({}) => {
